requirements.txt
fastapi
fastapi-sqlalchemy
psycopg2-binary
pydantic[dotenv]
uvicorn
alembic
SQLAlchemy
gunicorn
auth_lib_profcomff[fastapi]
